ollert_backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
psycopg2-binary==2.9.5
django-ordered-model==3.6